+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#2044](https://github.com/open-telemetry/opentelemetry-python/pull/2044))
 - `opentelemetry-api` Attribute keys must be non-empty strings.
   ([#2057](https://github.com/open-telemetry/opentelemetry-python/pull/2057))
+- Add ENV variables to pass ca bundle and client cert+key for zipkin exporter ([#2064](https://github.com/open-telemetry/opentelemetry-python/pull/2064)) 
 
 ## [0.23.1](https://github.com/open-telemetry/opentelemetry-python/pull/1987) - 2021-07-26